+++ android/app/src/main/java/com/trimixblender/barefootblender/MainActivity.java
@@ -1,0 +1,5 @@
+package com.trimixblender.barefootblender;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}