+++ src/main.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { Capacitor } from "@capacitor/core";
 import App from "./App";
 import "./index.css";
 
@@ -15,7 +16,7 @@ ReactDOM.createRoot(root).render(
   </React.StrictMode>
 );
 
-if ("serviceWorker" in navigator) {
+if ("serviceWorker" in navigator && !Capacitor.isNativePlatform()) {
   void import("virtual:pwa-register")
     .then(({ registerSW }) => {
       // eslint-disable-next-line @typescript-eslint/no-unsafe-call